+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Define Validity.entitiesFieldDef() and Validity.recordsFieldDef() new functions, bumping minor candidate version number
     - Define Validity.closestByRecords() new function
     - Define dtValidityDate new component
+    - Define Validity.englobingPeriodByRecords() new function
 
 ### 1.0.0